move convertCurrency
