+++ src/features/offerRefacto/helpers/offerCredit.ts
@@ -3,7 +3,7 @@ import { getOfferPrice } from 'features/offer/helpers/getOfferPrice/getOfferPric
 import { HasEnoughCreditType } from 'features/offerRefacto/types'
 import { UserProfile } from 'features/share/types'
 import { convertCentsToEuros } from 'libs/parsers/pricesConversion'
-import { RoundUnit, convertCurrency } from 'shared/currency/convertEuroToPacificFranc'
+import { RoundUnit, convertCurrency } from 'shared/currency/convertCurrency'
 import { Currency } from 'shared/currency/useGetCurrencyToDisplay'
 
 export const hasEnoughCredit = (
