+++ src/features/search/pages/modals/PriceModal/PriceModal.tsx
@@ -17,7 +17,7 @@ import { priceSchema } from 'features/search/helpers/schema/priceSchema/priceSch
 import { SearchState } from 'features/search/types'
 import { convertCentsToEuros } from 'libs/parsers/pricesConversion'
 import { usePacificFrancToEuroRate } from 'queries/settings/useSettings'
-import { RoundUnit, convertCurrency } from 'shared/currency/convertEuroToPacificFranc'
+import { RoundUnit, convertCurrency } from 'shared/currency/convertCurrency'
 import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
 import { Currency, useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { isCurrentBeneficiary } from 'shared/user/checkStatusType'
